+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Master Dev Branch
 
+### 2026-08-05
+- Warm MigrationStatus cache in Kernel
+
 ## Version v1.8.2 - Released: 2026-08-04
 
 ### 2026-08-04